Refactor for RSPEC & Added specs test fpr Player Class
