+++ lib/winner.rb
