+++ spec/main_spec.rb
@@ -1,0 +1,10 @@
+# spec/tic_tac_toe_spec.rb
+# frozen_string_literal: true
+
+# require '../bin/main.rb'
+
+describe '#main' do
+    describe "#validate_marker" do
+        
+    end
+end
